+++ src/dotnet/ReSharperPlugin.RimworldDev/ProjectTemplates/RimworldProjectTemplate/Source/MyRimWorldMod.cs
@@ -1,0 +1,10 @@
+using Verse;
+
+namespace MyRimWorldMod;
+
+public class MyRimWorldMod : Mod
+{
+        public MyRimWorldMod(ModContentPack content) : base(content)
+        {
+        }
+}