+++ src/rider/main/kotlin/run/RunState.kt
@@ -6,6 +6,7 @@ import com.intellij.execution.configurations.RunProfileState
 import com.intellij.execution.process.*
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ProgramRunner
+import com.intellij.ide.actions.searcheverywhere.evaluate
 import com.intellij.util.system.OS
 import com.jetbrains.rd.util.lifetime.Lifetime
 import com.jetbrains.rider.debugger.DebuggerWorkerProcessHandler
@@ -19,6 +20,8 @@ import kotlin.io.path.Path
 
 class RunState(
     private val rimworldLocation: String,
+    private val saveFilePath: String,
+    private val modListPath: String,
     private val rimworldState: RunProfileState,
     remoteConfiguration: RemoteConfiguration,
     executionEnvironment: ExecutionEnvironment,
@@ -60,13 +63,12 @@ class RunState(
         lifetime: Lifetime
     ): ExecutionResult {
         setupDoorstop()
-
         val result = super.execute(executor, runner, workerProcessHandler)
         ProcessTerminatedListener.attach(workerProcessHandler.debuggerWorkerRealHandler)
 
         val rimworldResult = rimworldState.execute(executor, runner)
         workerProcessHandler.debuggerWorkerRealHandler.addProcessListener(createProcessListener(rimworldResult?.processHandler))
-
+        
         return result
     }
 
@@ -77,6 +79,7 @@ class RunState(
                 processHandler.removeProcessListener(this)
 
                 siblingProcessHandler?.getProcess()?.destroy()
+                QuickStartUtils.tearDown(saveFilePath)
                 removeDoorstep()
             }
         }