+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Adding defName auto-completion in C# for DefDatabase and `[DefOf]`
  * When defining a RunConfig, you can now select a ModList config and Save file to be loaded for just this one launch
  * Added support for `[LoadAlias]`
+ * New Mod Template now adds a PublisherPlus configuration file
 
 ## 2023.4
  * Added support for Custom Def Classes